+++ packages/core/src/components/axes/toolbar.ts
@@ -4,7 +4,7 @@ import { Events, RenderTypes, ToolbarControlTypes } from '../../interfaces';
 import { Tools } from '../../tools';
 
 // D3 Imports
-import { event, select } from 'd3-selection';
+import { select } from 'd3-selection';
 
 export class Toolbar extends Component {
 	type = 'toolbar';
@@ -97,7 +97,7 @@ export class Toolbar extends Component {
 				select(this)
 					.select('button')
 					.on('click', d.clickFunction)
-					.on('keyup', () => {
+					.on('keyup', (event: KeyboardEvent) => {
 						if (
 							(event.key && event.key === 'Enter') ||
 							event.key === ' '
@@ -218,7 +218,7 @@ export class Toolbar extends Component {
 		}
 	}
 
-	toggleOverflowMenu() {
+	toggleOverflowMenu(event) {
 		if (this.isOverflowMenuOpen()) {
 			// hide overflow menu
 			this.updateOverflowMenu(false);
@@ -325,7 +325,7 @@ export class Toolbar extends Component {
 			iconSVGContent: `<circle cx="16" cy="8" r="2"></circle>
 							 <circle cx="16" cy="16" r="2"></circle>
 							 <circle cx="16" cy="24" r="2"></circle>`,
-			clickFunction: () => this.toggleOverflowMenu(),
+			clickFunction: (event) => this.toggleOverflowMenu(event),
 		};
 	}