+++ packages/core/src/components/graphs/bullet.ts
@@ -91,11 +91,13 @@ export class Bullet extends Component {
 			rangeBoxesEnter
 				.merge(rangeBoxes)
 				.attr('class', (d) => `range-box order-${d.order}`)
-				.transition(
-					this.services.transitions.getTransition(
-						'bullet-range-box-update-enter',
-						animate
-					)
+				.transition()
+				.call((t) =>
+					this.services.transitions.setupTransition({
+						transition: t,
+						name: 'bullet-range-box-update-enter',
+						animate,
+					})
 				)
 				.attr('d', (d, i) => {
 					/*
@@ -155,11 +157,13 @@ export class Bullet extends Component {
 			barsEnter
 				.merge(bars)
 				.classed('bar', true)
-				.transition(
-					this.services.transitions.getTransition(
-						'bullet-bar-update-enter',
-						animate
-					)
+				.transition()
+				.call((t) =>
+					this.services.transitions.setupTransition({
+						transition: t,
+						name: 'bullet-bar-update-enter',
+						animate,
+					})
 				)
 				.attr('class', (d) =>
 					this.model.getColorClassName({
@@ -218,11 +222,13 @@ export class Bullet extends Component {
 			linesEnter
 				.merge(lines)
 				.classed('marker', true)
-				.transition(
-					this.services.transitions.getTransition(
-						'bullet-marker-update-enter',
-						animate
-					)
+				.transition()
+				.call((t) =>
+					this.services.transitions.setupTransition({
+						transition: t,
+						name: 'bullet-marker-update-enter',
+						animate,
+					})
 				)
 				.attr('d', (d, i) => {
 					/*
@@ -283,11 +289,13 @@ export class Bullet extends Component {
 						d.value <= d.barValue ? 'over-bar' : ''
 					}`;
 				})
-				.transition(
-					this.services.transitions.getTransition(
-						'bullet-quartile-update-enter',
-						animate
-					)
+				.transition()
+				.call((t) =>
+					this.services.transitions.setupTransition({
+						transition: t,
+						name: 'bullet-quartile-update-enter',
+						animate,
+					})
 				)
 				.attr('d', ({ datum: d, value }, i) => {
 					/*
@@ -381,7 +389,7 @@ export class Bullet extends Component {
 
 		this.parent
 			.selectAll('path.bar')
-			.on('mouseover', function (datum) {
+			.on('mouseover', function (event, datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed('hovered', true);
 				hoveredElement.transition(
@@ -392,6 +400,7 @@ export class Bullet extends Component {
 
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEOVER, {
+					event,
 					element: hoveredElement,
 					datum,
 				});
@@ -406,6 +415,7 @@ export class Bullet extends Component {
 				);
 
 				self.services.events.dispatchEvent(Events.Tooltip.SHOW, {
+					event,
 					hoveredElement,
 					items: [
 						{
@@ -437,23 +447,27 @@ export class Bullet extends Component {
 					],
 				});
 			})
-			.on('mousemove', function (datum) {
+			.on('mousemove', function (event, datum) {
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEMOVE, {
+					event,
 					element: select(this),
 					datum,
 				});
 
-				self.services.events.dispatchEvent(Events.Tooltip.MOVE);
+				self.services.events.dispatchEvent(Events.Tooltip.MOVE, {
+					event,
+				});
 			})
-			.on('click', function (datum) {
+			.on('click', function (event, datum) {
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Bar.BAR_CLICK, {
+					event,
 					element: select(this),
 					datum,
 				});
 			})
-			.on('mouseout', function (datum) {
+			.on('mouseout', function (event, datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed('hovered', false);
 
@@ -465,6 +479,7 @@ export class Bullet extends Component {
 
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Bar.BAR_MOUSEOUT, {
+					event,
 					element: hoveredElement,
 					datum,
 				});