+++ packages/core/src/components/graphs/circle-pack.ts
@@ -5,7 +5,7 @@ import * as Configuration from '../../configuration';
 
 // D3 Imports
 import { hierarchy as d3Hierarchy, pack as D3Pack } from 'd3-hierarchy';
-import { event, select } from 'd3-selection';
+import { select } from 'd3-selection';
 
 import {
 	ColorClassNameTypes,
@@ -254,7 +254,7 @@ export class CirclePack extends Component {
 		const self = this;
 		this.parent
 			.selectAll('circle.node')
-			.on('mouseover', function (datum) {
+			.on('mouseover', function (event, datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed('hovered', true);
 
@@ -321,6 +321,7 @@ export class CirclePack extends Component {
 
 					// Show tooltip
 					self.services.events.dispatchEvent(Events.Tooltip.SHOW, {
+						event,
 						hoveredElement,
 						items: [
 							{
@@ -344,26 +345,30 @@ export class CirclePack extends Component {
 				self.services.events.dispatchEvent(
 					Events.CirclePack.CIRCLE_MOUSEOVER,
 					{
+						event,
 						element: hoveredElement,
 						datum,
 					}
 				);
 			})
-			.on('mousemove', function (datum) {
+			.on('mousemove', function (event, datum) {
 				const hoveredElement = select(this);
 
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(
 					Events.CirclePack.CIRCLE_MOUSEMOVE,
 					{
+						event,
 						element: hoveredElement,
 						datum,
 					}
 				);
 
-				self.services.events.dispatchEvent(Events.Tooltip.MOVE);
+				self.services.events.dispatchEvent(Events.Tooltip.MOVE, {
+					event,
+				});
 			})
-			.on('mouseout', function (datum) {
+			.on('mouseout', function (event, datum) {
 				const hoveredElement = select(this);
 				hoveredElement.classed('hovered', false);
 
@@ -375,6 +380,7 @@ export class CirclePack extends Component {
 				self.services.events.dispatchEvent(
 					Events.CirclePack.CIRCLE_MOUSEOUT,
 					{
+						event,
 						element: hoveredElement,
 						datum,
 					}
@@ -385,7 +391,7 @@ export class CirclePack extends Component {
 					hoveredElement,
 				});
 			})
-			.on('click', function (datum) {
+			.on('click', function (event, datum) {
 				const hoveredElement = select(this);
 				const disabled = hoveredElement.classed('non-focal');
 
@@ -437,6 +443,7 @@ export class CirclePack extends Component {
 				self.services.events.dispatchEvent(
 					Events.CirclePack.CIRCLE_CLICK,
 					{
+						event,
 						element: hoveredElement,
 						datum,
 					}