+++ packages/core/src/components/graphs/donut.ts
@@ -40,11 +40,13 @@ export class Donut extends Pie {
 			.style('font-size', () =>
 				options.donut.center.numberFontSize(radius)
 			)
-			.transition(
-				this.services.transitions.getTransition(
-					'donut-figure-enter-update',
-					animate
-				)
+			.transition()
+			.call((t) =>
+				this.services.transitions.setupTransition({
+					transition: t,
+					name: 'donut-figure-enter-update',
+					animate,
+				})
 			)
 			.tween('text', function () {
 				return self.centerNumberTween(select(this));