+++ packages/core/src/components/graphs/gauge.ts
@@ -427,34 +427,38 @@ export class Gauge extends Component {
 		const self = this;
 		this.parent
 			.selectAll('path.arc-foreground')
-			.on('mouseover', function (datum) {
+			.on('mouseover', function (event, datum) {
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Gauge.ARC_MOUSEOVER, {
+					event,
 					element: select(this),
 					datum,
 				});
 			})
-			.on('mousemove', function (datum) {
+			.on('mousemove', function (event, datum) {
 				const hoveredElement = select(this);
 
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Gauge.ARC_MOUSEMOVE, {
+					event,
 					element: hoveredElement,
 					datum,
 				});
 			})
-			.on('click', function (datum) {
+			.on('click', function (event, datum) {
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Gauge.ARC_CLICK, {
+					event,
 					element: select(this),
 					datum,
 				});
 			})
-			.on('mouseout', function (datum) {
+			.on('mouseout', function (event, datum) {
 				const hoveredElement = select(this);
 
 				// Dispatch mouse event
 				self.services.events.dispatchEvent(Events.Gauge.ARC_MOUSEOUT, {
+					event,
 					element: hoveredElement,
 					datum,
 				});