+++ packages/core/src/components/graphs/line.ts
@@ -139,11 +139,13 @@ export class Line extends Component {
 					.join(',');
 			})
 			// Transition
-			.transition(
-				this.services.transitions.getTransition(
-					'line-update-enter',
-					animate
-				)
+			.transition()
+			.call((t) =>
+				this.services.transitions.setupTransition({
+					transition: t,
+					name: 'line-update-enter',
+					animate,
+				})
 			)
 			.attr('opacity', (d) => (d.hidden ? 0 : 1))
 			.attr('d', (group) => {
@@ -179,12 +181,6 @@ export class Line extends Component {
 	};
 
 	destroy() {
-		// Remove event listeners
-		this.parent
-			.selectAll('path.line')
-			.on('mousemove', null)
-			.on('mouseout', null);
-
 		// Remove legend listeners
 		const eventsFragment = this.services.events;
 		eventsFragment.removeEventListener(