+++ packages/core/src/components/graphs/lollipop.ts
@@ -76,11 +76,13 @@ export class Lollipop extends Scatter {
 					originalClassName: 'line',
 				})
 			)
-			.transition(
-				this.services.transitions.getTransition(
-					'lollipop-line-update-enter',
-					animate
-				)
+			.transition()
+			.call((t) =>
+				this.services.transitions.setupTransition({
+					transition: t,
+					name: 'lollipop-line-update-enter',
+					animate,
+				})
 			)
 			.style('stroke', (d) =>
 				this.model.getFillColor(d[groupMapsTo], d[domainIdentifier], d)
@@ -178,12 +180,6 @@ export class Lollipop extends Scatter {
 	};
 
 	destroy() {
-		// Remove event listeners
-		this.parent
-			.selectAll('line.line')
-			.on('mousemove', null)
-			.on('mouseout', null);
-
 		// Remove legend listeners
 		const eventsFragment = this.services.events;
 		eventsFragment.removeEventListener(