+++ packages/core/src/model.ts
@@ -4,7 +4,6 @@ import { Tools } from './tools';
 import { Events, ScaleTypes, ColorClassNameTypes } from './interfaces';
 
 // D3
-import { map } from 'd3-collection';
 import { scaleOrdinal } from 'd3-scale';
 import { stack } from 'd3-shape';
 
@@ -210,13 +209,15 @@ export class ChartModel {
 		const { groupMapsTo } = options.data;
 		const displayData = this.getDisplayData(groups);
 
-		const stackKeys = map(displayData, (datum) => {
-			const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(
-				datum
-			);
+		const stackKeys = Tools.removeArrayDuplicates(
+			displayData.map((datum) => {
+				const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(
+					datum
+				);
 
-			return datum[domainIdentifier];
-		}).keys();
+				return datum[domainIdentifier].toString();
+			})
+		);
 
 		const axisPosition = this.services.cartesianScales.domainAxisPosition;
 		const scaleType = options.axes[axisPosition].scaleType;
@@ -245,6 +246,7 @@ export class ChartModel {
 					const domainIdentifier = this.services.cartesianScales.getDomainIdentifier(
 						datum
 					);
+
 					return (
 						datum[groupMapsTo] === dataGroupName &&
 						datum[domainIdentifier].toString() === key
@@ -258,6 +260,7 @@ export class ChartModel {
 					? correspondingDatum[rangeIdentifier]
 					: null;
 			});
+
 			return correspondingValues;
 		}) as any;
 	}
@@ -617,10 +620,9 @@ export class ChartModel {
 		const { ACTIVE, DISABLED } = Configuration.legend.items.status;
 		const options = this.getOptions();
 
-		const uniqueDataGroups = map(
-			data,
-			(datum) => datum[groupMapsTo]
-		).keys();
+		const uniqueDataGroups = Tools.removeArrayDuplicates(
+			data.map((datum) => datum[groupMapsTo])
+		);
 
 		// check if selectedGroups can be applied to chart with current data groups
 		if (options.data.selectedGroups.length) {