+++ packages/core/src/services/essentials/transitions.ts
@@ -7,9 +7,14 @@ import { Tools } from '../../tools';
 // D3 Imports
 import { Transition, transition } from 'd3-transition';
 
+interface setupTransitionConfigs {
+	transition?: any; // d3 types are causing issues here, hence why using `any`
+	name?: string;
+	animate?: boolean;
+}
+
 export class Transitions extends Service {
 	pendingTransitions = {};
-	// transitions: Transition<any, any, any, any>[];
 
 	init() {
 		this.services.events.addEventListener(Events.Model.UPDATE, () => {
@@ -18,7 +23,7 @@ export class Transitions extends Service {
 	}
 
 	getTransition(
-		name?: string,
+		name: string,
 		animate?: boolean
 	): Transition<any, any, any, any> {
 		if (this.model.getOptions().animations === false || animate === false) {
@@ -47,6 +52,22 @@ export class Transitions extends Service {
 		});
 
 		return t;
+	}
+
+	setupTransition({ transition: t, name, animate }: setupTransitionConfigs) {
+		this.pendingTransitions[t._id] = t;
+		t.on('end interrupt cancel', () => {
+			delete this.pendingTransitions[t._id];
+		});
+
+		if (animate !== true) {
+			return t.duration(0);
+		}
+
+		return t.duration(
+			Tools.getProperty(Configuration.transitions, name, 'duration') ||
+				Configuration.transitions.default.duration
+		);
 	}
 
 	getPendingTransitions() {