+++ packages/core/src/tools.ts
@@ -29,8 +29,12 @@ import {
 	DebounceSettings,
 } from 'lodash-es';
 
-import { mouse } from 'd3-selection';
-import { Numeric } from 'd3';
+// D3 Imports
+// @ts-ignore
+// ts-ignore is needed because `@types/d3`
+// is missing the `pointer` function
+import { pointer } from 'd3-selection';
+import { Numeric } from 'd3-array';
 
 // Functions
 export namespace Tools {
@@ -50,14 +54,14 @@ export namespace Tools {
 	export const fromPairs = lodashFromPairs;
 	export const some = lodashSome;
 
-	export function debounceWithD3MousePosition(fn, delay, element) {
+	export function debounceWithD3MousePosition(fn, delay, holder) {
 		var timer = null;
 		return function () {
 			const context = this;
 			const args = arguments;
 
 			//we get the D3 event here
-			context.mousePosition = mouse(element);
+			context.mousePosition = pointer(args[0], holder);
 
 			clearTimeout(timer);