+++ packages/core/stories/tutorials/getting-started/vanilla.ts
@@ -15,13 +15,13 @@ consistent markup, styles, and behavior in prototype and production work.
 
 **with yarn:**
 \`\`\`bash
-yarn add @carbon/charts d3@5.x
+yarn add @carbon/charts d3
 \`\`\`
 
 **with npm:**
 
 \`\`\`bash
-npm install --save @carbon/charts d3@5.x
+npm install --save @carbon/charts d3
 \`\`\`
 
 **Note:** you'd also need to install \`carbon-components\` if you're not using a bundled version of the library.